[FLINK-21151] Move asynchronous writing part out of RocksDB strategy
The previous preparations allow us to put this backend-independent part
into flink-runtime, where it can be used by other backends to write
snapshots in the same common format.

FullSnapshotResources is now the common thing that backends must
provide.
